+++ drivers/misc/mediatek/lens/mt6752/inc/LC898122AF.h
@@ -11,11 +11,11 @@
 //Structures
 typedef struct {
 //current position
-u32 u4CurrentPosition;
+unsigned long u4CurrentPosition;
 //macro position
-u32 u4MacroPosition;
+unsigned long u4MacroPosition;
 //Infiniti position
-u32 u4InfPosition;
+unsigned long u4InfPosition;
 //Motor Status
 bool          bIsMotorMoving;
 //Motor Open?
@@ -33,11 +33,11 @@ bool          bIsSupportSR;
 //H means "switch T and Q atomically"
 #define LC898122AFIOC_G_MOTORINFO _IOR(LC898122AF_MAGIC,0,stLC898122AF_MotorInfo)
 
-#define LC898122AFIOC_T_MOVETO _IOW(LC898122AF_MAGIC,1, u32)
+#define LC898122AFIOC_T_MOVETO _IOW(LC898122AF_MAGIC,1,unsigned long)
 
-#define LC898122AFIOC_T_SETINFPOS _IOW(LC898122AF_MAGIC,2, u32)
+#define LC898122AFIOC_T_SETINFPOS _IOW(LC898122AF_MAGIC,2,unsigned long)
 
-#define LC898122AFIOC_T_SETMACROPOS _IOW(LC898122AF_MAGIC,3, u32)
+#define LC898122AFIOC_T_SETMACROPOS _IOW(LC898122AF_MAGIC,3,unsigned long)
 
 #else
 #endif
